requirements.txt
wikipedia-api
python-dotenv
aiogram
